@@ -1,18 +1,16 @@
 global using Microsoft.AspNetCore.Components.Authorization;
-using Microsoft.AspNetCore.Components;
-using Microsoft.AspNetCore.Components.Web;
 using MudBlazor.Services;
 using Blazored.LocalStorage;
-using NTS.Client.Securities;
 using NTS.Client.Services.Contracts;
 using NTS.Client.Services;
+using YourApp.Client.Securities;
 
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
 builder.Services.AddScoped<IAuthService, AuthService>();
 builder.Services.AddScoped<CustomAuthenticationStateProvider>();
-builder.Services.AddScoped<AuthenticationStateProvider ,CustomAuthenticationStateProvider>();
+builder.Services.AddScoped<AuthenticationStateProvider, CustomAuthenticationStateProvider>();
 
 builder.Services.AddRazorPages();
 builder.Services.AddServerSideBlazor();
