@@ -1,16 +1,16 @@
-﻿using Microsoft.AspNetCore.Components.Authorization;
-using Blazored.LocalStorage;
+﻿using Blazored.LocalStorage;
 using System.Security.Claims;
 using System.Net.Http.Headers;
 using System.Text.Json;
 
-namespace NTS.Client.Securities
+namespace YourApp.Client.Securities
 {
     public class CustomAuthenticationStateProvider : AuthenticationStateProvider
     {
         private readonly HttpClient httpClient;
         private readonly ILocalStorageService localStorageService;
         private string? token;
+        private string? role;
 
         public CustomAuthenticationStateProvider(HttpClient httpClient, ILocalStorageService localStorageService)
         {
@@ -18,12 +18,14 @@ namespace NTS.Client.Securities
             this.localStorageService = localStorageService;
         }
 
+        // Get authentication state asynchronously
         public override async Task<AuthenticationState> GetAuthenticationStateAsync()
         {
             var identity = new ClaimsIdentity();
 
             try
             {
+                // Retrieve the token and role from localStorage
                 token = await localStorageService.GetItemAsStringAsync("Token");
 
                 if (!string.IsNullOrEmpty(token))
@@ -35,15 +37,42 @@ namespace NTS.Client.Securities
             }
             catch (Exception ex)
             {
-                Console.WriteLine($"Error During Authentication State Retrieval: {ex.Message}");
+                Console.WriteLine($"Error during authentication state retrieval: {ex.Message}");
             }
 
-            return new AuthenticationState(new ClaimsPrincipal(identity));
+            var claimsPrincipal = new ClaimsPrincipal(identity);
+
+            return new AuthenticationState(claimsPrincipal);
         }
 
+        // Parse JWT to extract claims
+        public static IEnumerable<Claim> ParseClaimsFromJwt(string jwt)
+        {
+            try
+            {
+                var parts = jwt.Split('.');
+                if (parts.Length != 3) throw new FormatException("Invalid JWT format");
 
-        // RefreshJwtToken need to Fix Tomorrow
+                var payload = parts[1];
+                var jsonBytes = Convert.FromBase64String(AddPadding(payload));
+                var claims = JsonSerializer.Deserialize<Dictionary<string, object>>(jsonBytes);
 
+                return claims?.Select(kvp => new Claim(kvp.Key, kvp.Value.ToString()!)) ?? Enumerable.Empty<Claim>();
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine($"Error parsing claims from JWT: {ex.Message}");
+                return Enumerable.Empty<Claim>();
+            }
+        }
+
+        // Add padding for Base64 string
+        private static string AddPadding(string base64)
+        {
+            return base64.PadRight(base64.Length + (4 - base64.Length % 4) % 4, '=');
+        }
+
+        // Refresh the authentication state
         public async Task RefreshAuthenticationStateAsync()
         {
             try
@@ -53,46 +82,8 @@ namespace NTS.Client.Securities
             }
             catch (Exception ex)
             {
-                Console.WriteLine($"Error Refreshing Authentication State: {ex.Message}");
+                Console.WriteLine($"Error refreshing authentication state: {ex.Message}");
             }
-        }
-
-        public static IEnumerable<Claim> ParseClaimsFromJwt(string jwt)
-        {
-            try
-            {
-                var parts = jwt.Split('.');
-                if (parts.Length != 3) throw new FormatException("Invalid JWT Format");
-
-                var payload = parts[1];
-                var jsonBytes = Convert.FromBase64String(AddPadding(payload));
-                var claims = JsonSerializer.Deserialize<Dictionary<string, object>>(jsonBytes);
-
-                var roleClaim = claims?.FirstOrDefault(c => c.Key == "role");
-                if (roleClaim.HasValue) // Check if the Key exists
-                {
-                    Console.WriteLine("Role: " + roleClaim.Value);
-                }
-
-                return claims?.Select(kvp => new Claim(kvp.Key, kvp.Value.ToString()!)) ?? Enumerable.Empty<Claim>();
-            }
-            catch (FormatException ex)
-            {
-                Console.WriteLine($"Invalid Token Format: {ex.Message}");
-                return Enumerable.Empty<Claim>();
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Error parsing claims from JWT: {ex.Message}");
-                return Enumerable.Empty<Claim>();
-            }
-        }
-
-
-
-        private static string AddPadding(string base64)
-        {
-            return base64.PadRight(base64.Length + (4 - base64.Length % 4) % 4, '=');
         }
     }
 }
