@@ -1,33 +1,30 @@
 using Microsoft.EntityFrameworkCore;
 using Microsoft.OpenApi.Models;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Swashbuckle.AspNetCore.Filters;
 using Microsoft.IdentityModel.Tokens;
 using System.Text;
 using NTS.Server.Database.DatabaseContext;
 using NTS.Server.Services;
 using NTS.Server.Services.Contracts;
-using NTS.Server.Domain.Models;
+using NTS.Server.Domain.DTOs;
+using Scalar.AspNetCore;
 
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
-
-// Dependency Injection
-builder.Services.AddScoped<IAuthService, AuthService>();
-builder.Services.Configure<EmailSettings>(builder.Configuration.GetSection("EmailSettings"));
-builder.Services.AddTransient<IEmailService, EmailService>();
-
-
-builder.Services.AddDbContext<ApplicationDbContext>(options =>
-    options.UseSqlServer(builder.Configuration.GetConnectionString("DefaultConnection")));
-
-
 builder.Services.AddControllers();
 
+// Database Connection
+builder.Services.AddDbContext<ApplicationDbContext>(options =>
+    options.UseSqlServer(builder.Configuration.GetConnectionString("DefaultConnection"))
+);
+
+
 // learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
-builder.Services.AddSwaggerGen(options => 
+// Configure Swagger/OpenAPI with JWT Bearer Authentication
+builder.Services.AddSwaggerGen(options =>
 {
+    // Add Security Definition for Bearer Token
     options.AddSecurityDefinition("Bearer", new OpenApiSecurityScheme
     {
         Description = "Standard Authorization header using the Bearer scheme (\"bearer {token}\")",
@@ -37,6 +34,7 @@ builder.Services.AddSwaggerGen(options =>
         Scheme = "Bearer"
     });
 
+    // Require the Bearer token for all API endpoints
     options.AddSecurityRequirement(new OpenApiSecurityRequirement
     {
         {
@@ -53,40 +51,45 @@ builder.Services.AddSwaggerGen(options =>
     });
 });
 
-var secretKey = builder.Configuration.GetValue<string>("AppSettings:Token");
+// Configuration CORS For Blazor Server App
+builder.Services.AddCors(policy =>
+{
+    policy.AddPolicy("AllowBlazorApp", policy =>
+    {
+        policy
+        .WithOrigins("https://localhost:5000", "http://localhost:5001")
+        .AllowAnyHeader()
+        .AllowAnyMethod();
+    });
+});
 
-// JWT Authentication
+// JWT Bearer Authentication
 builder.Services.AddAuthentication(JwtBearerDefaults.AuthenticationScheme)
     .AddJwtBearer(options =>
     {
         options.TokenValidationParameters = new TokenValidationParameters
         {
-            ValidateIssuerSigningKey = true,
-            IssuerSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(secretKey!)),
-            ValidateIssuer = false,
-            ValidateAudience = false,
+            ValidateIssuer = true,
+            ValidIssuer = builder.Configuration["AppSettings:Issuer"],
+            ValidateAudience = true,
+            ValidAudience = builder.Configuration["AppSettings:Audience"],
             ValidateLifetime = true,
-            ClockSkew = TimeSpan.FromMinutes(5)
+            IssuerSigningKey = new SymmetricSecurityKey(
+                Encoding.UTF8.GetBytes(builder.Configuration["AppSettings:Token"]!)),
+            ValidateIssuerSigningKey = true
         };
     });
 
+// Dependency Injection for services
+builder.Services.AddScoped<IAuthService, AuthService>();
+builder.Services.Configure<EmailSettingsDto>(builder.Configuration.GetSection("EmailSettings"));
+builder.Services.AddTransient<IEmailService, EmailService>();
 
-builder.Services.AddCors(policy =>
-{
-    policy.AddPolicy("AllowBlazorApp",
-        policy =>
-        {
-            policy
-            .WithOrigins("https://localhost:5000","http://localhost:5001")
-            .AllowAnyHeader()
-            .AllowAnyMethod();
-        });
-});
-
+builder.Services.AddOpenApi();
 
 var app = builder.Build();
 
-// Configure the HTTP request pipeline.
+// Configure the HTTP request pipeline in development
 if (app.Environment.IsDevelopment())
 {
     app.UseSwagger();
@@ -95,8 +98,10 @@ if (app.Environment.IsDevelopment())
 
 app.UseHttpsRedirection();
 app.UseCors("AllowBlazorApp");
+
 app.UseAuthentication();
 app.UseAuthorization();
+
 app.MapControllers();
 
 app.Run();
